Rename package to extendedsecuritysettings
