+++ src/main/resources/io/jenkins/plugins/extendedsecuritysettings/DisableLoginAutocompleteDecorator/simple-footer.jelly
