+++ src/main/resources/io/jenkins/plugins/extendedsecuritysettings/ParanoiaConfiguration/config.jelly
