+++ src/main/resources/io/jenkins/plugins/extendedsecuritysettings/ParanoiaConfiguration/help-enableXssProtectionHeader.html
